+++ src/gui/crystalSettingsWidget.h
@@ -3,23 +3,31 @@
 #include <QGroupBox>
 #include <QComboBox>
 #include <QLabel>
+#include <QPushButton>
+#include <QSpinBox>
+#include <QDataWidgetMapper>
 #include "sliderSpinBox.h"
+#include "crystalModel.h"
 #include "../simulation/crystalPopulation.h"
+#include "../simulation/crystalPopulationRepository.h"
 
 class CrystalSettingsWidget : public QGroupBox
 {
     Q_OBJECT
 public:
-    CrystalSettingsWidget(QWidget *parent = nullptr);
-
-    void SetInitialValues(HaloSim::CrystalPopulation crystal);
+    CrystalSettingsWidget(std::shared_ptr<HaloSim::CrystalPopulationRepository> crystalRepository, QWidget *parent = nullptr);
 
 signals:
-    void crystalChanged(HaloSim::CrystalPopulation crystal);
+    void crystalChanged();
 
 private:
     void setupUi();
-    HaloSim::CrystalPopulation stateToCrystalPopulation() const;
+    SliderSpinBox *createAngleSlider(double min, double max);
+    void setTiltVisibility(bool);
+    void setRotationVisibility(bool);
+
+    QPushButton *mAddPopulationButton;
+    QPushButton *mRemovePopulationButton;
 
     SliderSpinBox *mCaRatioSlider;
     SliderSpinBox *mCaRatioStdSlider;
@@ -35,4 +43,12 @@ private:
     QLabel *mRotationAverageLabel;
     SliderSpinBox *mRotationStdSlider;
     QLabel *mRotationStdLabel;
+
+    QSpinBox *mWeightSpinBox;
+
+    CrystalModel *mModel;
+    QDataWidgetMapper *mMapper;
+    QComboBox *mPopulationComboBox;
+
+    unsigned int mNextPopulationId;
 };