+++ src/gui/sliderSpinBox.h
@@ -7,8 +7,10 @@
 class SliderSpinBox : public QWidget
 {
     Q_OBJECT
+    Q_PROPERTY(double value READ value WRITE setValue USER true)
 public:
     SliderSpinBox(QWidget *parent = nullptr);
+    SliderSpinBox(double min, double max, QWidget *parent = nullptr);
     void setSuffix(const QString &suffix);
     double value() const;