+++ src/gui/viewSettingsWidget.cpp
@@ -6,7 +6,12 @@ ViewSettingsWidget::ViewSettingsWidget(QWidget *parent)
 {
     setupUi();
 
-    auto cameraChangeHandler = [this]() { cameraChanged(stateToCamera()); };
+    auto cameraChangeHandler = [this]() {
+        auto camera = stateToCamera();
+        mFieldOfViewSlider->setMaximum(camera.getMaximumFov());
+        camera.fov = mFieldOfViewSlider->value();
+        cameraChanged(camera);
+    };
 
     connect(mCameraProjectionComboBox, QOverload<int>::of(&QComboBox::currentIndexChanged), cameraChangeHandler);
     connect(mFieldOfViewSlider, &SliderSpinBox::valueChanged, cameraChangeHandler);
@@ -19,6 +24,8 @@ ViewSettingsWidget::ViewSettingsWidget(QWidget *parent)
 
 void ViewSettingsWidget::setupUi()
 {
+    setMaximumWidth(400);
+
     mCameraProjectionComboBox = new QComboBox();
     mCameraProjectionComboBox->addItems({"Stereographic",
                                          "Rectilinear",
@@ -38,8 +45,9 @@ void ViewSettingsWidget::setupUi()
     mYawSlider->setSuffix("°");
 
     mFieldOfViewSlider = new SliderSpinBox();
-    mFieldOfViewSlider->setMinimum(0.01);
-    mFieldOfViewSlider->setMaximum(2.0);
+    mFieldOfViewSlider->setMinimum(10.0);
+    mFieldOfViewSlider->setMaximum(360.0);
+    mFieldOfViewSlider->setSuffix("°");
 
     mBrightnessSlider = new SliderSpinBox();
     mBrightnessSlider->setMinimum(0.1);