+++ src/simulation/simulationEngine.cpp
@@ -15,7 +15,8 @@ namespace HaloSim
 
 SimulationEngine::SimulationEngine(
     unsigned int outputWidth,
-    unsigned int outputHeight)
+    unsigned int outputHeight,
+    std::shared_ptr<CrystalPopulationRepository> crystalRepository)
     : mOutputWidth(outputWidth),
       mOutputHeight(outputHeight),
       mMersenneTwister(std::mt19937(std::random_device()())),
@@ -26,8 +27,8 @@ SimulationEngine::SimulationEngine(
       mInitialized(false),
       mCamera(DefaultCamera()),
       mLight(DefaultLightSource()),
-      mCrystals(DefaultCrystalPopulation()),
-      mCameraLockedToLightSource(false)
+      mCameraLockedToLightSource(false),
+      mCrystalRepository(crystalRepository)
 {
 }
 
@@ -41,7 +42,7 @@ Camera SimulationEngine::GetCamera() const
     return mCamera;
 }
 
-void SimulationEngine::SetCamera(const struct Camera camera)
+void SimulationEngine::SetCamera(const Camera camera)
 {
     Clear();
     mCamera = camera;
@@ -49,17 +50,6 @@ void SimulationEngine::SetCamera(const struct Camera camera)
     {
         PointCameraToLightSource();
     }
-}
-
-CrystalPopulation SimulationEngine::GetCrystalPopulation() const
-{
-    return mCrystals;
-}
-
-void SimulationEngine::SetCrystalPopulation(const CrystalPopulation crystals)
-{
-    Clear();
-    mCrystals = crystals;
 }
 
 LightSource SimulationEngine::GetLightSource() const
@@ -112,34 +102,43 @@ void SimulationEngine::Step()
 
     mSimulationShader->bind();
 
-    unsigned int seed = mUniformDistribution(mMersenneTwister);
+    for (auto i = 0u; i < mCrystalRepository->GetCount(); ++i)
+    {
+        glMemoryBarrier(GL_SHADER_IMAGE_ACCESS_BARRIER_BIT);
+        unsigned int seed = mUniformDistribution(mMersenneTwister);
 
-    /* The following line needs to use glUniform1ui instead of the
-       setUniformValue method because of a bug in Qt:
-       https://bugreports.qt.io/browse/QTBUG-45507
-    */
-    glUniform1ui(glGetUniformLocation(mSimulationShader->programId(), "rngSeed"), seed);
-    mSimulationShader->setUniformValue("sun.altitude", mLight.altitude);
-    mSimulationShader->setUniformValue("sun.diameter", mLight.diameter);
+        auto crystals = mCrystalRepository->Get(i);
+        auto probability = mCrystalRepository->GetProbability(i);
+        auto numRays = static_cast<unsigned int>(mRaysPerStep * probability);
 
-    mSimulationShader->setUniformValue("crystalProperties.caRatioAverage", mCrystals.caRatioAverage);
-    mSimulationShader->setUniformValue("crystalProperties.caRatioStd", mCrystals.caRatioStd);
+        /*
+        The following line needs to use glUniform1ui instead of the
+        setUniformValue method because of a bug in Qt:
+        https://bugreports.qt.io/browse/QTBUG-45507
+        */
+        glUniform1ui(glGetUniformLocation(mSimulationShader->programId(), "rngSeed"), seed);
+        mSimulationShader->setUniformValue("sun.altitude", mLight.altitude);
+        mSimulationShader->setUniformValue("sun.diameter", mLight.diameter);
 
-    mSimulationShader->setUniformValue("crystalProperties.tiltDistribution", mCrystals.tiltDistribution);
-    mSimulationShader->setUniformValue("crystalProperties.tiltAverage", mCrystals.tiltAverage);
-    mSimulationShader->setUniformValue("crystalProperties.tiltStd", mCrystals.tiltStd);
+        mSimulationShader->setUniformValue("crystalProperties.caRatioAverage", crystals.caRatioAverage);
+        mSimulationShader->setUniformValue("crystalProperties.caRatioStd", crystals.caRatioStd);
 
-    mSimulationShader->setUniformValue("crystalProperties.rotationDistribution", mCrystals.rotationDistribution);
-    mSimulationShader->setUniformValue("crystalProperties.rotationAverage", mCrystals.rotationAverage);
-    mSimulationShader->setUniformValue("crystalProperties.rotationStd", mCrystals.rotationStd);
+        mSimulationShader->setUniformValue("crystalProperties.tiltDistribution", crystals.tiltDistribution);
+        mSimulationShader->setUniformValue("crystalProperties.tiltAverage", crystals.tiltAverage);
+        mSimulationShader->setUniformValue("crystalProperties.tiltStd", crystals.tiltStd);
 
-    mSimulationShader->setUniformValue("camera.pitch", mCamera.pitch);
-    mSimulationShader->setUniformValue("camera.yaw", mCamera.yaw);
-    mSimulationShader->setUniformValue("camera.fov", mCamera.fov);
-    mSimulationShader->setUniformValue("camera.projection", mCamera.projection);
-    mSimulationShader->setUniformValue("camera.hideSubHorizon", mCamera.hideSubHorizon ? 1 : 0);
+        mSimulationShader->setUniformValue("crystalProperties.rotationDistribution", crystals.rotationDistribution);
+        mSimulationShader->setUniformValue("crystalProperties.rotationAverage", crystals.rotationAverage);
+        mSimulationShader->setUniformValue("crystalProperties.rotationStd", crystals.rotationStd);
 
-    glDispatchCompute(mRaysPerStep, 1, 1);
+        mSimulationShader->setUniformValue("camera.pitch", mCamera.pitch);
+        mSimulationShader->setUniformValue("camera.yaw", mCamera.yaw);
+        mSimulationShader->setUniformValue("camera.fov", mCamera.fov);
+        mSimulationShader->setUniformValue("camera.projection", mCamera.projection);
+        mSimulationShader->setUniformValue("camera.hideSubHorizon", mCamera.hideSubHorizon ? 1 : 0);
+
+        glDispatchCompute(numRays, 1, 1);
+    }
 }
 
 void SimulationEngine::Clear()